Update display for fractional hours

+++ app/src/main/java/tech/jpco/nazztimesheets/model/WorkLog.kt
@@ -4,7 +4,7 @@ package tech.jpco.nazztimesheets.model
  * Created by Dave - Work on 10/15/2017.
  */
 class WorkLog : ArrayList<WorkSession>() {
-    fun getAheadState(fractionDenominator: Int): Float {
+    fun getAheadState(fractionDenominator: Int): Double {
         val minionAhead = fold(0, { minionAhead, next ->
             minionAhead + next.minutesElapsed() *
                     (if (next.type == WorkSession.WorkType.PERSONAL) -1 else 1)
@@ -13,8 +13,8 @@ class WorkLog : ArrayList<WorkSession>() {
         return minutesToFractionalHours(minionAhead, fractionDenominator)
     }
 
-    private fun minutesToFractionalHours(minutes: Int, denom: Int): Float {
+    private fun minutesToFractionalHours(minutes: Int, denom: Int): Double {
         return (minutes / 60) +
-                (((minutes % 60) * denom) / 60).toFloat() / denom
+                (((minutes % 60) * denom) / 60).toDouble() / denom
     }
 }
